requirements.txt

pytz
Flask
pandas
vectorbt
psycopg2
sqlalchemy
python-dotenv
python-binance
